feat(runtime): extract /actions and /mcp dispatcher domains — ADR-0076 D11 step ③ PR-9 (#2462)
The deep-coupling batch rides the PR-8 subsystem out:

- domains/actions.ts: ADR-0066 D4 gate + ADR-0104 param contract on
  actionExec.*; the env pre-resolution block becomes two deps seams —
  getDefaultEnvironmentId + resolveProjectKernelObjectQL (the ADR-0006
  Phase-5 direct-caller kernel swap; the this.kernel write side effect
  stays dispatcher-owned). The redundant leading/trailing-slash regex
  (CodeQL redos twin flagged in #2462) drops for split+filter.
- domains/mcp.ts: transport + /mcp/skill + OAuth resource-metadata +
  the principal-bound bridge; legacy /mcp/skill-before-/mcp precedence
  reproduced with ordered entries incl. '?' forms. buildMcpBridge kept
  as a public thin delegate for direct callers (tests).
- Statics gotcha round 2: HttpDispatcher.isMcpEnabled — extracted body
  referenced the class static; inlined to isMcpServerEnabled() at both
  sites and the wrapper dropped.
- The authz-conformance identity pin for buildMcpBridge(context)
  follows the body: discover file/re/key now point at domains/mcp.ts.

Verified: runtime 649, http-conformance 41, dogfood 351, full build
green.
